Changed datatype of flags and ingredients from long to int since the expected number of values <32 mio

+++ ANMAIServer/src/foodServer/Flag.java
@@ -6,16 +6,16 @@ package foodServer;
  *
  */
 public class Flag implements IFlag{
-	private long id;
+	private int id;
 	private String name;
 	private String description;
 	private boolean limitToProduct;
 
-	public long getId() {
+	public int getId() {
 		return this.id;
 	}
 
-	public void setId(long aId) {
+	public void setId(int aId) {
 		this.id = aId;
 	}
